Clear the list of lambdas, so that subsequent units don't get zombies

+++ Gemini/Compiler.cpp
@@ -2281,7 +2281,7 @@ void Compiler::GenerateSentinel()
     size_t paddingSize = alignedSize - mCodeBin.size();
     size_t curIndex = ReserveProgram( paddingSize );
 
-    for ( int i = 0; i < paddingSize; i++ )
+    for ( size_t i = 0; i < paddingSize; i++ )
     {
         mCodeBin[curIndex++] = OP_SENTINEL;
     }
